change namespaces and improve invoke tests
