+++ src/OpenRiaServices.DomainServices.Client.Web/Framework/Data/Resource.resx
@@ -117,17 +117,8 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="DomainClient_UnableToDetermineHostUri" xml:space="preserve">
-    <value>The current application or application host is null and the host URI cannot be determined.</value>
-  </data>
-  <data name="DomainContext_InvalidServiceUri" xml:space="preserve">
+  <data name="WebDomainClient_InvalidServiceUri" xml:space="preserve">
     <value>The specified service URI is not correctly formatted.</value>
-  </data>
-  <data name="DomainContext_LoadOperationFailed" xml:space="preserve">
-    <value>Load operation failed for query '{0}'. {1}</value>
-  </data>
-  <data name="DomainContext_SubmitOperationFailed" xml:space="preserve">
-    <value>Submit operation failed. {0}</value>
   </data>
   <data name="PoxBinaryMessageEncoder_InvalidContentType" xml:space="preserve">
     <value>The PoxBinaryMessageEncoder only supports content type {0}.</value>
