+++ src/OpenRiaServices.DomainServices.Client/Test/Client.Test/Data/OperationTests.cs
@@ -9,6 +9,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using DataTests.Northwind.LTS;
 using System.ComponentModel.DataAnnotations;
 using OpenRiaServices.Silverlight.Testing;
+using System.Collections;
 
 namespace OpenRiaServices.DomainServices.Client.Test
 {
@@ -131,16 +132,6 @@ namespace OpenRiaServices.DomainServices.Client.Test
             // verify the exception properties
             Assert.AreSame(ex, expectedException);
 
-            // TODO: Add separate test with mock DomainClient which throws a specific exception
-            // and move the following checks there
-            /*
-            Assert.IsNotNull(expectedException);
-            Assert.AreEqual(string.Format(Resource.DomainContext_LoadOperationFailed, "ThrowGeneralException", ex.Message), expectedException.Message);
-            Assert.AreEqual(ex.StackTrace, expectedException.StackTrace);
-            Assert.AreEqual(ex.Status, expectedException.Status);
-            Assert.AreEqual(ex.ErrorCode, expectedException.ErrorCode);
-            */
-
             Assert.AreEqual(false, lo.IsErrorHandled);
 
             // now test again with validation errors
@@ -160,7 +151,7 @@ namespace OpenRiaServices.DomainServices.Client.Test
 
             // verify the exception properties
             Assert.AreSame(expectedException, ex);;
-            CollectionAssert.AreEqual(validationErrors, lo.ValidationErrors.ToList());
+            CollectionAssert.AreEqual(validationErrors, (ICollection)lo.ValidationErrors);
         }
 
         /// <summary>
@@ -186,23 +177,18 @@ namespace OpenRiaServices.DomainServices.Client.Test
             }
 
             // verify the exception properties
-            Assert.IsNotNull(expectedException);
-            Assert.AreEqual(string.Format(Resource.DomainContext_InvokeOperationFailed, "Echo", ex.Message), expectedException.Message);
-            Assert.AreEqual(ex.StackTrace, expectedException.StackTrace);
-            Assert.AreEqual(ex.Status, expectedException.Status);
-            Assert.AreEqual(ex.ErrorCode, expectedException.ErrorCode);
-
+            Assert.AreSame(ex, expectedException);
             Assert.AreEqual(false, invoke.IsErrorHandled);
 
             // now test again with validation errors
             expectedException = null;
             ValidationResult[] validationErrors = new ValidationResult[] { new ValidationResult("Foo", new string[] { "Bar" }) };
             invoke = new InvokeOperation("Echo", null, null, null, false);
+            var validationException = new DomainOperationException("validation", validationErrors);
 
             try
             {
-                // TODO: Rewrite
-                invoke.Complete(new DomainOperationException("", validationErrors));
+                invoke.Complete(validationException);
             }
             catch (DomainOperationException e)
             {
@@ -210,8 +196,8 @@ namespace OpenRiaServices.DomainServices.Client.Test
             }
 
             // verify the exception properties
-            Assert.IsNotNull(expectedException);
-            Assert.AreEqual(string.Format(Resource.DomainContext_InvokeOperationFailed_Validation, "Echo"), expectedException.Message);
+            Assert.AreSame(validationException, expectedException);
+            CollectionAssert.AreEqual(validationErrors, (ICollection)invoke.ValidationErrors);
         }
 
         /// <summary>
