+++ src/OpenRiaServices.DomainServices.Client/Test/Client.Test/Data/PoxTests.cs
@@ -6,8 +6,9 @@ using System.ServiceModel.Channels;
 using System.Xml;
 using System.Xml.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using OpenRiaServices.DomainServices.Client.Test;
 
-namespace OpenRiaServices.DomainServices.Client.Test
+namespace OpenRiaServices.DomainServices.Client.Web.Test
 {
     [TestClass]
     public class PoxTests
