+++ src/OpenRiaServices.DomainServices.Hosting/Framework/Services/MessageEncoders/BufferManagerStream.cs
@@ -5,7 +5,7 @@ using System.ServiceModel.Channels;
 #if SERVERFX
 namespace OpenRiaServices.DomainServices.Hosting
 #else
-namespace OpenRiaServices.DomainServices.Client
+namespace OpenRiaServices.DomainServices.Client.Web
 #endif
 {
     /// <summary>
