+++ src/Test/Desktop/OpenRiaServices.Common.DomainServices.Test/Mocks/ServerSideAsyncDomainService.cs
@@ -37,12 +37,12 @@ namespace TestDomainServices
 
         private Task Delay(int milliseconds)
         {
-            return Task.Factory.StartNew(() => Thread.Sleep(milliseconds));
+            return Task.Delay(milliseconds);
         }
 
         private Task Delay(TimeSpan delay)
         {
-            return Task.Factory.StartNew(() => Thread.Sleep(delay));
+            return Task.Delay(delay);
         }
 
         public IQueryable<RangeItem> GetRange()
